Wraper Builder de RigidBodies creado, ajuste de ejemplo de triangulos
