+++ TGC.Examples/Bullet/Physics/PhysicsGame.cs
@@ -47,39 +47,5 @@ namespace TGC.Examples.Bullet.Physics
 
         public abstract int getElements();
 
-        //TODO Wrapper Builder bodys.
-        public RigidBody CreateBall(float size, float mass, float originX, float originY, float originZ)
-        {
-            //Crea una bola de radio 10 origen 50 de 1 kg.
-            var ballShape = new SphereShape(size);
-            var ballTransform = TGCMatrix.Identity;
-            ballTransform.Origin = new TGCVector3(originX, originY, originZ);
-            var ballMotionState = new DefaultMotionState(ballTransform.ToBsMatrix);
-            //Podriamos no calcular la inercia para que no rote, pero es correcto que rote tambien.
-            var ballLocalInertia = ballShape.CalculateLocalInertia(mass);
-            var ballInfo = new RigidBodyConstructionInfo(1, ballMotionState, ballShape, ballLocalInertia);
-            var ballBody = new RigidBody(ballInfo);
-            ballBody.LinearFactor = TGCVector3.One.ToBsVector;
-            ballBody.SetDamping(0.1f, 0.5f);
-            ballBody.Restitution = 0.5f;
-            return ballBody;
-        }
-
-        public RigidBody CreateBox(float size, float mass, float x, float y, float z, float yaw, float pitch, float roll)
-        {
-            //Se crea una caja de tamaño 20 con rotaciones y origien en 10,100,10 y 1kg de masa.
-            var boxShape = new BoxShape(size, size, size);
-            var boxTransform = TGCMatrix.RotationYawPitchRoll(yaw, pitch, roll).ToBsMatrix;
-            boxTransform.Origin = new TGCVector3(x, y, z).ToBsVector;
-            DefaultMotionState boxMotionState = new DefaultMotionState(boxTransform);
-            //Es importante calcular la inercia caso contrario el objeto no rotara.
-            var boxLocalInertia = boxShape.CalculateLocalInertia(mass);
-            var boxInfo = new RigidBodyConstructionInfo(1f, boxMotionState, boxShape, boxLocalInertia);
-            var boxBody = new RigidBody(boxInfo);
-            boxBody.LinearFactor = TGCVector3.One.ToBsVector;
-            //boxBody.SetDamping(0.7f, 0.9f);
-            //boxBody.Restitution = 1f;
-            return boxBody;
-        }
     }
 }
