+++ TGC.Examples/Bullet/Physics/WallBullet.cs
@@ -50,7 +50,7 @@ namespace TGC.Examples.Bullet.Physics
             {
                 for (var j = 0; j < 10; j++)
                 {
-                    var boxBody = this.CreateBox(10f, 1, i * 20f + 5f, j * 20f + 5f, 0f, 0, 0, 0);
+                    var boxBody = Core.BulletPhysics.BulletRigidBodyConstructor.CreateBox(new TGCVector3(10,10,10),1,new TGCVector3(i * 20f + 5f, j * 20f + 5f, 0f),0,0,0,0.5f);
                     boxBodys.Add(boxBody);
                     dynamicsWorld.AddRigidBody(boxBody);
                 }
@@ -83,39 +83,39 @@ namespace TGC.Examples.Bullet.Physics
 
             if (Ctx.Input.keyUp(Key.Space))
             {
-                var ballBody = this.CreateBall(10f, 1f, Ctx.Camara.Position.X, Ctx.Camara.Position.Y, Ctx.Camara.Position.Z);
+                var ballBody = Core.BulletPhysics.BulletRigidBodyConstructor.CreateBall(10f, 1f, new TGCVector3(Ctx.Camara.Position.X, Ctx.Camara.Position.Y, Ctx.Camara.Position.Z));
                 var dir = new TGCVector3(Ctx.Camara.LookAt.X - Ctx.Camara.Position.X, Ctx.Camara.LookAt.Y - Ctx.Camara.Position.Y, Ctx.Camara.LookAt.Z - Ctx.Camara.Position.Z).ToBsVector;
                 dir.Normalize();
                 ballBody.LinearVelocity = dir * 900;
                 ballBody.LinearFactor = TGCVector3.One.ToBsVector;
-                ballBody.SetDamping(0.1f, 0.5f);
-                ballBody.Restitution = 0.5f;
+                /*ballBody.SetDamping(0.1f, 0.5f);
+                ballBody.Restitution = 0.5f;*/
                 ballBodys.Add(ballBody);
                 dynamicsWorld.AddRigidBody(ballBody);
             }
 
             if (Ctx.Input.keyUp(Key.Q))
             {
-                var ballBody = this.CreateBall(10f, 10f, Ctx.Camara.Position.X, Ctx.Camara.Position.Y, Ctx.Camara.Position.Z);
+                var ballBody = Core.BulletPhysics.BulletRigidBodyConstructor.CreateBall(10f, 10f, new TGCVector3(Ctx.Camara.Position.X, Ctx.Camara.Position.Y, Ctx.Camara.Position.Z));
                 var dir = new TGCVector3(Ctx.Camara.LookAt.X - Ctx.Camara.Position.X, Ctx.Camara.LookAt.Y - Ctx.Camara.Position.Y, Ctx.Camara.LookAt.Z - Ctx.Camara.Position.Z).ToBsVector;
                 dir.Normalize();
                 ballBody.LinearVelocity = dir * 900;
                 ballBody.LinearFactor = TGCVector3.One.ToBsVector;
-                ballBody.SetDamping(0.1f, 0.5f);
-                ballBody.Restitution = 0.5f;
+                /*ballBody.SetDamping(0.1f, 0.5f);
+                ballBody.Restitution = 0.5f;*/
                 ballBodys.Add(ballBody);
                 dynamicsWorld.AddRigidBody(ballBody);
             }
 
             if (Ctx.Input.keyUp(Key.W))
             {
-                var ballBody = this.CreateBall(10f, 0.1f, Ctx.Camara.Position.X, Ctx.Camara.Position.Y, Ctx.Camara.Position.Z);
+                var ballBody = Core.BulletPhysics.BulletRigidBodyConstructor.CreateBall(10f, 0.1f, new TGCVector3( Ctx.Camara.Position.X, Ctx.Camara.Position.Y, Ctx.Camara.Position.Z));
                 var dir = new TGCVector3(Ctx.Camara.LookAt.X - Ctx.Camara.Position.X, Ctx.Camara.LookAt.Y - Ctx.Camara.Position.Y, Ctx.Camara.LookAt.Z - Ctx.Camara.Position.Z).ToBsVector;
                 dir.Normalize();
                 ballBody.LinearVelocity = dir * 900;
                 ballBody.LinearFactor = TGCVector3.One.ToBsVector;
-                ballBody.SetDamping(0.1f, 0.1f);
-                ballBody.Restitution = 0.9f;
+                /*ballBody.SetDamping(0.1f, 0.1f);
+                ballBody.Restitution = 0.9f;*/
                 ballBodys.Add(ballBody);
                 dynamicsWorld.AddRigidBody(ballBody);
             }
