+++ TGC.Examples/TGC.Examples.csproj
@@ -96,7 +96,6 @@
     <Compile Include="Bullet\Physics\HelloWorldBullet.cs" />
     <Compile Include="Bullet\Physics\HelloWorldBullet2.cs" />
     <Compile Include="Bullet\Physics\PhysicsGame.cs" />
-    <Compile Include="Bullet\Physics\PhysicsHelper.cs" />
     <Compile Include="Bullet\Physics\TrianglePhysics.cs" />
     <Compile Include="Bullet\Physics\WallBullet.cs" />
     <Compile Include="Camara\TgcFpsCamera.cs" />
